Added edit method to User as well as some setters, Removed ability to edit UserName from Profile
In my opinion, it makes no sense for a user to be able to edit his
username from Edit Profile, and it would potentially cause a lot of
headaches with checking if he changed his username to another user or if
he kept it the same, etc. So I removed that ability.

Clicking "Save" will crash Edit Profile, but bradshawz is working on
fixing the saving over network in non-main thread issue, so will wait
for that before applying fix.