Updated Thing Constructor, removed borrowing rate, started edit game
Thing constructor now keeps track of number of players and category,
which might be useful later on for future searches with more
functionality.
I removed borrowing rating from Add a Game because I think that's an
added complication to the bids.

TODO: Make save button on Edit a Game and Edit Profile work