Added internal topBidAmount to thing which is checked each time you add a bid.
This way, we dont need to do an O(n) check of all bids every time to
calculate the top bid amount.
We might still need to do some async stuff in GetBidFragment if we
choose to have a list of offers there, but according to the TA, we onyl
need the top bid amount, so we can always leave that until later.